+++ examples/manage_container_server/main.go
@@ -13,57 +13,14 @@ TODO:
 package main
 
 import (
-	"errors"
 	"flag"
 	"fmt"
-	"html/template"
 	"log"
 	"net/http"
 	"os"
-	"strconv"
-	"strings"
-
-	"github.com/DmitryBogomolov/containerator"
-	"github.com/DmitryBogomolov/containerator/manage"
-	"github.com/docker/docker/client"
 )
 
 const defaultPort = 4001
-
-var errBadURL = errors.New("bad url")
-var errNoProject = errors.New("no project")
-
-type projectItem struct {
-	Name       string
-	ConfigPath string
-}
-
-type projectsCache struct {
-	Dir   string
-	Items []projectItem
-}
-
-func (obj *projectsCache) refresh() {
-	obj.Items = []projectItem{
-		projectItem{
-			Name:       "Project 1",
-			ConfigPath: "/at",
-		},
-		projectItem{
-			Name:       "Project 2",
-			ConfigPath: "/gv",
-		},
-	}
-}
-
-func (obj *projectsCache) get(name string) *projectItem {
-	for i, item := range obj.Items {
-		if item.Name == name {
-			return &obj.Items[i]
-		}
-	}
-	return nil
-}
 
 func checkHTTPMethod(method string, w http.ResponseWriter, r *http.Request) bool {
 	if r.Method != method {
@@ -73,103 +30,8 @@ func checkHTTPMethod(method string, w http.ResponseWriter, r *http.Request) bool
 	return true
 }
 
-type rootHandler struct {
-	cache    *projectsCache
-	template *template.Template
-}
-
-func newRootHandler(cache *projectsCache) (*rootHandler, error) {
-	tmpl := template.Must(template.ParseFiles("page.html"))
-	return &rootHandler{
-		cache:    cache,
-		template: tmpl,
-	}, nil
-}
-
-func (h *rootHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if !checkHTTPMethod(http.MethodGet, w, r) {
-		return
-	}
-	h.template.Execute(w, h.cache)
-}
-
-type commandHandler struct {
-	cache *projectsCache
-	cli   client.CommonAPIClient
-}
-
-func newCommandHandler(cache *projectsCache) (*commandHandler, error) {
-	cli, err := client.NewEnvClient()
-	if err != nil {
-		return nil, err
-	}
-	return &commandHandler{
-		cache: cache,
-		cli:   cli,
-	}, nil
-}
-
-func (h *commandHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if !checkHTTPMethod(http.MethodPost, w, r) {
-		return
-	}
-	name := strings.Replace(r.URL.Path, "/manage/", "", 1)
-	item := h.cache.get(name)
-	if item == nil {
-		http.Error(w, fmt.Sprintf("'%s' is not found\n", name), http.StatusNotFound)
-		return
-	}
-	body, err := invokeManage(h.cli, item.ConfigPath, r)
-	if err != nil {
-		http.Error(w, fmt.Sprintf("Error: %v\n", err), http.StatusInternalServerError)
-		return
-	}
-	w.WriteHeader(http.StatusOK)
-	fmt.Fprintln(w, body)
-}
-
-// func findTarget(workDir string, name string) (string, error) {
-// 	items, err := filepath.Glob(filepath.Join(workDir, name, "*.yaml"))
-// 	if err == nil && len(items) == 0 {
-// 		err = errNoProject
-// 	}
-// 	if err != nil {
-// 		return "", err
-// 	}
-// 	return items[0], nil
-// }
-
-func parseBool(value string) bool {
-	ret, _ := strconv.ParseBool(value)
-	return ret
-}
-
-func invokeManage(cli client.CommonAPIClient, configPath string, r *http.Request) (string, error) {
-	err := r.ParseForm()
-	if err != nil {
-		return "", err
-	}
-	options := &manage.Options{
-		Mode:   r.PostFormValue("mode"),
-		Tag:    r.PostFormValue("tag"),
-		Remove: parseBool(r.PostFormValue("remove")),
-		Force:  parseBool(r.PostFormValue("force")),
-	}
-	config, err := manage.ReadConfig(configPath)
-	if err != nil {
-		return "", err
-	}
-	cont, err := manage.Manage(cli, config, options)
-	if err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("Container: %s", containerator.GetContainerName(cont)), nil
-}
-
 func setupServer() (http.Handler, error) {
-	workDir, _ := os.Getwd()
-	cache := &projectsCache{Dir: workDir}
-	cache.refresh()
+	cache := newProjectsCache()
 	rootHandler, err := newRootHandler(cache)
 	if err != nil {
 		return nil, err
